pyproject.toml
[tool.poetry]
name = "pdf-text-extraction"
version = "0.1.0"
description = ""
authors = ["Michael Polonskij <michael.polonskij@gmx.de>"]
readme = "README.md"
packages = [{include = "pdf_text_extraction"}]

[tool.poetry.dependencies]
python = "^3.11"
pypdf2 = "^3.0.1"
pdfminer-six = "^20221105"
pdfplumber = "^0.10.3"
pillow = "^10.1.0"
pdf2image = "^1.16.3"
pytesseract = "^0.3.10"
pyprojroot = "^0.3.0"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
